+++ app/src/main/java/com/example/hangouts/homeScreen/hangoutCreation/HangoutLocationSelectionMapFragment.java
@@ -1,6 +1,5 @@
-package com.example.hangouts.homeScreen.fragments;
+package com.example.hangouts.homeScreen.hangoutCreation;
 
-import androidx.activity.result.ActivityResultCallback;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
@@ -20,12 +19,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.Toast;
 
 import com.example.hangouts.BuildConfig;
 import com.example.hangouts.R;
 import com.example.hangouts.databinding.FragmentHangoutLocationSelectionMapBinding;
-import com.example.hangouts.loginScreen.SignupFragment;
+import com.example.hangouts.homeScreen.utils.LocationUtils;
 import com.google.android.gms.common.api.Status;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
