+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/homeFragmentContainer"
-        android:name="com.example.hangouts.homeScreen.fragments.HomeFragment"
+        android:name="com.example.hangouts.homeScreen.HomeFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
