+++ app/src/main/res/layout/fragment_hangout_location_selection_map.xml
@@ -38,7 +38,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/autocompleteFragment"
-            tools:context=".homeScreen.fragments.HangoutLocationSelectionMapFragment" />
+            tools:context=".homeScreen.hangoutCreation.HangoutLocationSelectionMapFragment" />
 
         <fragment
             android:id="@+id/autocompleteFragment"
